Cometi un crimen

+++ files/main.js
@@ -1,21 +1,24 @@
-let f = document.getElementById("forma");
-let n = document.getElementById("nombre");
-let i = document.getElementById("ingresar");
-if (localStorage["nombre"]) {
+let forma = document.getElementById("forma");
+let nombre = document.getElementById("nombre");
+let contraseña = document.getElementById("contraseña");
+if (
+  localStorage["nombre"] &&
+  localStorage["contraseña"] &&
+  //TODO: Reformular esto, que use una funcion especialmente para saber si el nombre y contraseña coinciden
+  !verificar_usuario(localStorage["nombre"], localStorage["contraseña"])
+) {
   document.getElementById("ingresar").textContent =
-    "Hola " + localStorage["nombre"] + ", Quieres cambiar de nombre?";
-  f.hidden = true;
+    "Hola " + localStorage["nombre"] + ", Quieres cambiar de usuario?";
+  forma.hidden = true;
   let s = document.createElement("button");
-  s.setAttribute("id", "Si");
   s.appendChild(document.createTextNode("Si"));
   document.body.appendChild(s);
   let no = document.createElement("button");
-  no.setAttribute("id", "No");
   no.appendChild(document.createTextNode("No"));
   document.body.appendChild(no);
   s.onclick = function () {
     localStorage.clear();
-    f.hidden = false;
+    forma.hidden = false;
     s.hidden = true;
     no.hidden = true;
   };
@@ -23,10 +26,38 @@ if (localStorage["nombre"]) {
     document.location.href = "http://127.0.0.1:8000/lobby";
   };
 }
-f.addEventListener("submit", function (a) {
-  if (!n.value) {
+forma.addEventListener("submit", function (a) {
+  a.preventDefault();
+  if (!nombre.value) {
     alert("El nombre esta vacio");
-    a.preventDefault();
+    return;
   }
-  localStorage["nombre"] = n.value;
+  if (!contraseña.value) {
+    alert("La contraseña esta vacia");
+    return;
+  }
+
+  if (verificar_usuario(nombre.value, contraseña.value)) {
+    localStorage["nombre"] = nombre.value;
+    localStorage["contraseña"] = contraseña.value;
+    document.location.href = "http://127.0.0.1:8000/lobby";
+  } else {
+    alert("Contraseña equivocada (o algo malo paso)");
+  }
 });
+
+async function verificar_usuario(nombre, contraseña) {
+  let lobbys = ["main"];
+  let user = await fetch("/registrarse", {
+    method: "POST",
+    body: new URLSearchParams({ nombre, contraseña, lobbys }),
+  })
+    .then((response) => response.json())
+    .then((data) => {
+      console.log(data);
+      if (data && data.code == 202) {
+        return true;
+      }
+      return false;
+    });
+}
